content-agent/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
requests==2.32.5
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose==3.5.0
cryptography==46.0.3
itsdangerous==2.2.0
apscheduler==3.11.1
python-dotenv==1.2.1